Removed TypeDef
