+++ Assets/SpacetimeDB/Scripts/Identity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 924c91782f80dffba9114a596343ef14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
