--- Assets/SpacetimeDB/Scripts/TypeDef.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 6ebc8a2d7cebaea6e9e6bd87e94781bd
-timeCreated: 1657552127
